syntax: dismantle ast_util.
